+++ web/modules/custom/itkdev_booking/assets/src/components/create-booking-filters.js
@@ -91,6 +91,34 @@ function CreateBookingFilters({
     }
 
     setResourceCategoryOptions(newResourceCategoryOptions);
+
+    // Add filtering based on filterParams
+    if (filterParams) {
+      const filterLocations = filterParams["location[]"].reduce((carry, locationParam) => {
+        const foundLocations = locations.filter((aLocation) => aLocation.value === locationParam);
+
+        foundLocations.forEach((foundLocation) => carry.push(foundLocation));
+
+        return carry;
+      }, []);
+
+      setLocationFilter(filterLocations);
+
+      const filterResourceMails = filterParams["resourceMail[]"].reduce((carry, resourceParam) => {
+        const foundResources = allResources.filter((resource) => resource.resourceMail === resourceParam);
+
+        foundResources.forEach((foundResource) =>
+          carry.push({
+            value: foundResource.resourceMail,
+            label: foundResource.resourceDisplayName ?? foundResource.resourceName,
+          })
+        );
+
+        return carry;
+      }, []);
+
+      setResourceFilter(filterResourceMails);
+    }
   }, [allResources]);
 
   // Set location filter and resource dropdown options.
@@ -311,7 +339,10 @@ CreateBookingFilters.defaultProps = {
 };
 
 CreateBookingFilters.propTypes = {
-  filterParams: PropTypes.shape({}).isRequired,
+  filterParams: PropTypes.shape({
+    "location[]": PropTypes.arrayOf(PropTypes.string),
+    "resourceMail[]": PropTypes.arrayOf(PropTypes.string),
+  }).isRequired,
   setFilterParams: PropTypes.func.isRequired,
   allResources: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
   disabled: PropTypes.bool.isRequired,